% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadWInstall.R
\name{loadWInstall}
\alias{loadWInstall}
\title{Function loadWInstall() - loading packages smartly}
\usage{
loadWInstall(...)
}
\arguments{
\item{...}{packages to be loaded}
}
\value{
nothing
}
\description{
Function loadWInstall() - loading packages smartly
}
\examples{
loadWInstall(ggplot2, dplyr, lubridate)
}

